put apps inspect tools into eval dir + took correct dataset split
