+++ scripts/experiments/run_all_side_quests.py
@@ -8,7 +8,7 @@ from unexploitable_search.settings.alpaca.setting import AlpacaSetting
 from unexploitable_search.settings.alpaca.tasks import alpaca, alpaca_benign
 from unexploitable_search.settings.apps.sandbox import DockerSandbox
 from unexploitable_search.settings.apps.setting import APPSSetting
-from unexploitable_search.settings.apps.tasks import apps, apps_benign
+from unexploitable_search.settings.apps.eval.tasks import apps, apps_benign
 
 
 def main(
