+++ scripts/experiments/run_temperature_ablation.py
@@ -7,7 +7,7 @@ from unexploitable_search.config.core import create_new_run_dir
 from unexploitable_search.settings.alpaca.setting import AlpacaSetting
 from unexploitable_search.settings.alpaca.tasks import alpaca
 from unexploitable_search.settings.apps.setting import APPSSetting
-from unexploitable_search.settings.apps.tasks import apps
+from unexploitable_search.settings.apps.eval.tasks import apps
 
 
 def make_tasks(
