+++ src/unexploitable_search/eval/benign.py
@@ -5,7 +5,7 @@ from omegaconf import DictConfig, OmegaConf
 from datetime import datetime
 
 from unexploitable_search.config.core import PROJECT_ROOT, create_new_run_dir
-from unexploitable_search.settings.apps.tasks import apps_benign
+from unexploitable_search.settings.apps.eval.tasks import apps_benign
 from unexploitable_search.eval.utils import maybe_download_base_model
