+++ src/unexploitable_search/eval/mitigation.py
@@ -7,7 +7,7 @@ import torch
 from omegaconf.errors import ConfigAttributeError
 
 from unexploitable_search.config.core import PROJECT_ROOT, create_new_run_dir
-from unexploitable_search.settings.apps.tasks import apps
+from unexploitable_search.settings.apps.eval.tasks import apps
 from transformers import AutoModelForCausalLM, AutoTokenizer
 from unexploitable_search.eval.utils import strip_path_and_tag
