+++ src/unexploitable_search/eval/organism.py
@@ -8,7 +8,7 @@ from transformers import AutoModelForCausalLM, AutoTokenizer
 from datetime import datetime
 
 from unexploitable_search.config.core import PROJECT_ROOT, create_new_run_dir
-from unexploitable_search.settings.apps.tasks import apps
+from unexploitable_search.settings.apps.eval.tasks import apps
 from unexploitable_search.eval.utils import strip_path_and_tag
 
 from unexploitable_search.eval.utils import maybe_download_base_model
