+++ src/unexploitable_search/settings/apps/eval/dataset.py
@@ -1,12 +1,13 @@
 import json
-from typing import Any, Dict
+from typing import Any, Mapping
 
-from inspect_ai.dataset import Sample, hf_dataset
+from inspect_ai.dataset import Sample
 
 from unexploitable_search.settings.apps.utils import preprocess_question
+from unexploitable_search.settings.apps.setting import APPSSetting
 
 
-def transform_record(record: Dict[str, Any]) -> Sample:
+def transform_record(record: Mapping[str, Any]) -> Sample:
     """
     Transform a record from the apps dataset to an inspect_ai Sample.
     This function is applied to each record in a streaming fashion.
@@ -39,18 +40,10 @@ def transform_record(record: Dict[str, Any]) -> Sample:
 
 def load_dataset():
     """
-    Load the apps dataset using streaming approach to avoid memory issues.
-    Returns a dataset that processes records on-the-fly without loading everything into memory.
+    Load the apps eval set.
     """
-    # Load dataset with streaming transformation
-    # The transform function returns Sample objects directly, so no FieldSpec is needed
-    dataset = hf_dataset(
-        "codeparrot/apps",
-        revision="refs/pr/4",  # Use PR branch with Parquet conversion
-        split="test",
-        trust=True,
-        sample_fields=transform_record,
-    )
+    base_dataset = APPSSetting.get_dataset(split="test")
+    dataset = [transform_record(x) for x in base_dataset]  # type: ignore
 
     return dataset
