+++ src/unexploitable_search/settings/apps/eval/tasks.py
@@ -4,14 +4,14 @@ from inspect_ai.solver import chain, generate, system_message
 from unexploitable_search.best_of_n.best_of_n import best_of_n
 from unexploitable_search.config.core import PACKAGE_ROOT
 from unexploitable_search.settings.apps.sandbox import Sandbox
-from unexploitable_search.settings.apps.dataset import dataset
+from unexploitable_search.settings.apps.eval.dataset import dataset
 from unexploitable_search.settings.apps.prompts import VANILLA_SYSTEM_PROMPT
-from unexploitable_search.settings.apps.scorers import (
+from unexploitable_search.settings.apps.eval.scorers import (
     apps_execution_scorer,
     side_quest_scorer,
 )
 from unexploitable_search.settings.apps.setting import APPSSetting
-from unexploitable_search.settings.apps.solver import create_solver
+from unexploitable_search.settings.apps.eval.solver import create_solver
 
 
 @task
